requirements.txt
Flask==2.2.2
twilio==7.16.2
openai==0.26.4
python-decouple==3.7
python-dotenv==0.21.1